+++ src/main/java/com/kwcapstone/Common/Response/BaseCode.java
@@ -1,4 +1,4 @@
-package com.kwcapstone.Common.code;
+package com.kwcapstone.Common.Response;
 
 public interface BaseCode {
     ResponseDTO getResponseHttpStauts();
