+++ src/main/java/com/kwcapstone/Common/Response/BaseErrorResponse.java
@@ -1,9 +1,7 @@
-package com.kwcapstone.Common;
+package com.kwcapstone.Common.Response;
 
 import com.kwcapstone.Exception.BaseException;
 import lombok.Getter;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.server.ResponseStatusException;
 
 @Getter
 public class BaseErrorResponse extends BaseResponse {
