+++ src/main/java/com/kwcapstone/Common/Response/ResponseDTO.java
@@ -1,10 +1,8 @@
-package com.kwcapstone.Common.code;
+package com.kwcapstone.Common.Response;
 
-import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
-import org.springframework.stereotype.Service;
 
 @Getter
 @Builder
