+++ src/main/java/com/kwcapstone/Service/GoogleOAuthService.java
@@ -3,11 +3,8 @@ package com.kwcapstone.Service;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.kwcapstone.Common.BaseResponse;
 import com.kwcapstone.Domain.Entity.Member;
-import com.kwcapstone.Exception.BaseException;
 import com.kwcapstone.GoogleLogin.Auth.GoogleUser;
-import com.kwcapstone.Naver.Dto.NaverProfile;
 import com.kwcapstone.Repository.MemberRepository;
 import com.kwcapstone.Token.Domain.Token;
 import com.kwcapstone.Token.JwtTokenProvider;
@@ -15,7 +12,6 @@ import com.kwcapstone.Token.Repository.TokenRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.PropertySource;
-import org.springframework.core.ParameterizedTypeReference;
 import org.springframework.stereotype.Service;
 import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
@@ -25,9 +21,7 @@ import org.springframework.http.*;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.net.URLDecoder;
-import java.net.URLEncoder;
 import java.nio.charset.StandardCharsets;
-import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
