+++ src/main/java/com/kwcapstone/Token/Controller/TokenController.java
@@ -1,14 +1,12 @@
 package com.kwcapstone.Token.Controller;
 
-import com.kwcapstone.Common.BaseResponse;
-import com.kwcapstone.Common.code.SuccessStatus;
+import com.kwcapstone.Common.Response.BaseResponse;
+import com.kwcapstone.Common.Response.SuccessStatus;
 import com.kwcapstone.Token.Domain.Dto.TokenRequest;
 import com.kwcapstone.Token.Domain.Dto.TokenResponse;
 import com.kwcapstone.Token.Service.TokenService;
 import io.swagger.v3.oas.annotations.Operation;
-import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
